feat: add quantity-based unit billing with product unit selection
Implement weight/volume/quantity unit handling in products and billing,
including unit conversion, decimal qty for weight/volume, and iOS 15 SPM
migration. Update app icons and branding assets.

+++ test/utils/helpers/bill_tax_helper_test.dart
@@ -4,7 +4,7 @@ import 'package:thuga/utils/helpers/bill_tax_helper.dart';
 
 CartItemModel _item({
   required double price,
-  int quantity = 1,
+  double quantity = 1,
   double sgst = 0,
   double cgst = 0,
 }) {
